Rename contract package to notary
Weird name, since it does not contain any contract logic.
