+++ auth/services/notary/notary_test.go
@@ -16,7 +16,7 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package contract
+package notary
 
 import (
 	"context"
